chore: move logout method to kratos
